Add: MemoListItem 구조 변경
