--- src/components/molecule/etc/Option.js
@@ -1,5 +1,0 @@
-const Option = ({ value, name }) => {
-  return <option value={value}>{name}</option>;
-};
-
-export default Option;
