+++ src/components/organisms/box/SelectBox.js
@@ -1,31 +1,49 @@
+import React from "react";
 import styled from "styled-components";
-import Option from "../../molecule/etc/Option";
+import { useNavigate } from "react-router-dom";
+import { sortOptionList, filterOptionList } from "../../../util/optionList";
+import Button from "../../molecule/etc/Button";
+import Select from "../../molecule/etc/Select";
 
-const SelectBox = ({ value, onChange, optionList }) => {
+const SelectBox = ({ sortType, setSortType, filter, setFilter }) => {
+  const navigate = useNavigate();
+  const goNew = () => {
+    navigate("/new");
+  };
   return (
-    <ControlMenu value={value} onChange={(e) => onChange(e.target.value)}>
-      {optionList &&
-        optionList.map((list, index) => (
-          <Option key={index} value={list.value} name={list.name} />
-        ))}
-    </ControlMenu>
+    <Wrapper>
+      <LeftCol>
+        <Select
+          value={sortType}
+          onChange={setSortType}
+          optionList={sortOptionList}
+        />
+        <Select
+          value={filter}
+          onChange={setFilter}
+          optionList={filterOptionList}
+        />
+      </LeftCol>
+      <RightCol>
+        <Button type={"positive"} text={"새로운 기록 남기기"} onClick={goNew} />
+      </RightCol>
+    </Wrapper>
   );
 };
 
-const ControlMenu = styled.select`
-  margin-right: 10px;
-  border: none;
-  border-radius: 5px;
-  background-color: #ececec;
+const Wrapper = styled.div`
+  margin-top: 20px;
+  margin-bottom: 30px;
 
-  padding-top: 10px;
-  padding-bottom: 10px;
-  padding-left: 20px;
-  padding-right: 20px;
-
-  cursor: pointer;
-  /* font-family: "Nanum Pen Script", cursive; */
-  font-size: 18px;
+  display: flex;
+  justify-content: space-between;
+`;
+const LeftCol = styled.div``;
+const RightCol = styled.div`
+  flex-grow: 1;
+  & > button {
+    width: 100%;
+  }
 `;
 
 export default SelectBox;
