+++ src/components/template/ViewContainer.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useState, useMemo } from "react";
 import { useSelector } from "react-redux";
 import { getMonthDate } from "../../util/date";
 import Button from "../molecule/etc/Button";
